+++ utility/install_python.sh
@@ -3,7 +3,7 @@
 echo "***********************************************"
 echo "Installing Python 3.6"
 echo "***********************************************"
-add-apt-repository -y ppa:jonathonf/python-3.6
+add-apt-repository -y ppa:deadsnakes/ppa
 apt-get -y update
 apt-get -y install python3.6
 apt-get -y install python3.6-dev